Decouple `grace-core` from `grace-plugin-domain-class`
- Update tests
- Move tests from `grace-core` into `grace-plugin-domain-class`
