+++ grace-core/build.gradle
@@ -15,7 +15,6 @@ dependencies {
     compileOnly libs.spring.tx
 
     testImplementation project(":grace-test")
-    testImplementation project(":grace-plugin-domain-class")
     testImplementation libs.jakarta.servlet
     testImplementation libs.spring.jdbc
     testImplementation libs.spring.test
