+++ grace-plugin-domain-class/src/test/groovy/org/grails/core/artefact/DomainClassArtefactHandlerSpec.groovy
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2022-2025 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.grails.core.artefact
+
+import org.codehaus.groovy.ast.ClassNode
+import spock.lang.Specification
+
+import grails.core.ArtefactHandler
+
+/**
+ * @author Michael Yan
+ * @since 2024.0.0
+ */
+class DomainClassArtefactHandlerSpec extends Specification {
+
+    void "Check TestEntity annotated with GORM '@Entity'"() {
+        given:
+        ArtefactHandler handler = new DomainClassArtefactHandler()
+        GroovyClassLoader gcl = new GroovyClassLoader()
+        Class<?> clazz = gcl.parseClass('''
+@grails.gorm.annotation.Entity
+class TestEntity {
+}
+''')
+
+        ClassNode classNode = new ClassNode(clazz)
+
+        expect:
+        handler.isArtefact(classNode)
+        handler.isArtefact(clazz)
+    }
+
+}
